refactor(app): move ProtocolUpload hooks
These hooks are widely used, including in the ODD, and thus shouldn't
live in the desktop-specific ProtocolUpload organism. Also, move some deps.
