+++ app/src/organisms/Devices/ProtocolRun/ProtocolRunHeader/RunHeaderBannerContainer/TerminalRunBannerContainer.tsx
@@ -13,9 +13,9 @@ import {
 import { RUN_STATUS_STOPPED, RUN_STATUS_SUCCEEDED } from '@opentrons/api-client'
 
 import { Banner } from '/app/atoms/Banner'
-import { useCloseCurrentRun } from '../../../../ProtocolUpload/hooks'
+import { useCloseCurrentRun } from '/app/resources/runs'
 import { useIsRunCurrent } from '/app/resources/runs'
-import { useMostRecentRunId } from '../../../../ProtocolUpload/hooks/useMostRecentRunId'
+import { useMostRecentRunId } from '/app/resources/runs/useMostRecentRunId'
 
 import type { RunHeaderBannerContainerProps } from '.'
