+++ app/src/organisms/Devices/ProtocolRun/ProtocolRunHeader/RunHeaderContent/RunHeaderSectionLower.tsx
@@ -17,7 +17,7 @@ import { RUN_STATUS_RUNNING } from '@opentrons/api-client'
 import { formatTimestamp } from '../../../utils'
 import { useRunControls } from '../../../../RunTimeControl/hooks'
 import { EMPTY_TIMESTAMP, useRunTimestamps } from '/app/resources/runs'
-import { useCloseCurrentRun } from '../../../../ProtocolUpload/hooks'
+import { useCloseCurrentRun } from '/app/resources/runs'
 import { LabeledValue } from './LabeledValue'
 import { isCancellableStatus } from '../utils'
