+++ app/src/organisms/Devices/utils.ts
@@ -9,10 +9,7 @@ import type {
   Instruments,
   PipetteData,
   PipetteOffsetCalibration,
-  RunTimeParameterFilesCreateData,
-  RunTimeParameterValuesCreateData,
 } from '@opentrons/api-client'
-import type { RunTimeParameter } from '@opentrons/shared-data'
 
 /**
  * formats a string if it is in ISO 8601 date format
@@ -92,46 +89,4 @@ export function getShowPipetteCalibrationWarning(
       } else return false
     }) ?? false
   )
-}
-
-/**
- * prepares object to send to endpoints requiring RunTimeParameterValuesCreateData
- * @param {RunTimeParameter[]} runTimeParameters array of updated RunTimeParameter overrides
- * @returns {RunTimeParameterValuesCreateData} object mapping variable name to value
- */
-export function getRunTimeParameterValuesForRun(
-  runTimeParameters: RunTimeParameter[]
-): RunTimeParameterValuesCreateData {
-  return runTimeParameters.reduce((acc, param) => {
-    const { variableName } = param
-    if (param.type !== 'csv_file' && param.value !== param.default) {
-      return { ...acc, [variableName]: param.value }
-    }
-    return acc
-  }, {})
-}
-
-/**
- * prepares object to send to endpoints requiring RunTimeParameterFilesCreateData
- * @param {RunTimeParameter[]} runTimeParameters array of updated RunTimeParameter overrides
- * @param {Record<string, string>} [fileIdMap] mapping of variable name to file ID created and returned by robot server
- * @returns {RunTimeParameterFilesCreateData} object mapping variable name to file ID
- */
-export function getRunTimeParameterFilesForRun(
-  runTimeParameters: RunTimeParameter[],
-  fileIdMap?: Record<string, string>
-): RunTimeParameterFilesCreateData {
-  return runTimeParameters.reduce((acc, param) => {
-    const { variableName } = param
-    if (param.type === 'csv_file' && param.file?.id != null) {
-      return { ...acc, [variableName]: param.file.id }
-    } else if (
-      param.type === 'csv_file' &&
-      fileIdMap != null &&
-      variableName in fileIdMap
-    ) {
-      return { ...acc, [variableName]: fileIdMap[variableName] }
-    }
-    return acc
-  }, {})
 }
