+++ app/src/organisms/DropTipWizardFlows/__tests__/TipsAttachedModal.test.tsx
@@ -9,7 +9,7 @@ import { i18n } from '/app/i18n'
 import { handleTipsAttachedModal } from '../TipsAttachedModal'
 import { LEFT } from '@opentrons/shared-data'
 import { mockPipetteInfo } from '/app/redux/pipettes/__fixtures__'
-import { useCloseCurrentRun } from '../../ProtocolUpload/hooks'
+import { useCloseCurrentRun } from '/app/resources/runs'
 import { useDropTipWizardFlows } from '..'
 
 import type { Mock } from 'vitest'
@@ -17,7 +17,7 @@ import type { PipetteModelSpecs } from '@opentrons/shared-data'
 import type { HostConfig } from '@opentrons/api-client'
 import type { PipetteWithTip } from '../hooks'
 
-vi.mock('../../ProtocolUpload/hooks')
+vi.mock('/app/resources/runs/useCloseCurrentRun')
 vi.mock('..')
 
 const MOCK_ACTUAL_PIPETTE = {
