--- app/src/organisms/ProtocolUpload/hooks/index.ts
@@ -1,4 +1,0 @@
-export * from './useCloseCurrentRun'
-export * from './useCurrentRunCommands'
-export * from './useCloneRun'
-export * from './useRestartRun'
