+++ app/src/pages/ODD/RunSummary/index.tsx
@@ -42,7 +42,6 @@ import {
 } from '@opentrons/react-api-client'
 import { useRunControls } from '/app/organisms/RunTimeControl/hooks'
 import { useRunCreatedAtTimestamp } from '/app/organisms/Devices/hooks'
-import { useCloseCurrentRun } from '/app/organisms/ProtocolUpload/hooks'
 import { onDeviceDisplayFormatTimestamp } from '/app/organisms/Devices/utils'
 import { RunTimer } from '/app/organisms/Devices/ProtocolRun/RunTimer'
 import {
@@ -62,6 +61,7 @@ import {
   useIsRunCurrent,
   useNotifyRunQuery,
   useRunTimestamps,
+  useCloseCurrentRun,
   EMPTY_TIMESTAMP,
 } from '/app/resources/runs'
 import {
