+++ app/src/resources/runs/index.ts
@@ -11,4 +11,9 @@ export * from './useCurrentRun'
 export * from './useRunTimestamps'
 export * from './useRunCommands'
 export * from './useRunStatus'
+export * from './useCloneRun'
+export * from './useCloseCurrentRun'
+export * from './useCurrentRunCommands'
+export * from './useMostRecentRunId'
+export * from './useRestartRun'
 export * from './constants'
