+++ app/src/resources/runs/useCloneRun.ts
@@ -5,11 +5,11 @@ import {
   useCreateRunMutation,
   useCreateProtocolAnalysisMutation,
 } from '@opentrons/react-api-client'
-import { useNotifyRunQuery } from '/app/resources/runs'
+import { useNotifyRunQuery } from './useNotifyRunQuery'
 import {
   getRunTimeParameterValuesForRun,
   getRunTimeParameterFilesForRun,
-} from '../../Devices/utils'
+} from '/app/transformations/runs'
 
 import type { Run } from '@opentrons/api-client'
