+++ app/src/resources/runs/useCloseCurrentRun.ts
